backend/requirements.txt
# 📦 Updated Requirements for Email + 2FA
# DATEI: backend/requirements.txt

# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Redis & Caching
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
cryptography==41.0.7

# Email & Validation
email-validator==2.1.0
aiosmtplib==3.0.1

# 🔐 2FA & QR Code Generation
pyotp==2.9.0
qrcode[pil]==7.4.2

# Data Validation
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Rate Limiting
slowapi==0.1.9

# Environment Variables
python-decouple==3.8